Moved profile_helpers into its subpackage
